+++ Tabby/src/Platforms/MacOS/MacOSInput.cpp
@@ -1,4 +1,4 @@
-#ifdef  TB_PLATFORM_MACOS
+#ifdef TB_PLATFORM_MACOS
 
 #include "Tabby/Core/Input.h"
 
@@ -7,38 +7,38 @@
 
 namespace Tabby {
 
-	bool Input::IsKeyPressed(const KeyCode key)
-	{
-		auto* window = static_cast<GLFWwindow*>(Application::Get().GetWindow().GetNativeWindow());
-		auto state = glfwGetKey(window, static_cast<int32_t>(key));
-		return state == GLFW_PRESS;
-	}
+bool Input::IsKeyPressed(const KeyCode key)
+{
+    auto* window = static_cast<GLFWwindow*>(Application::Get().GetWindow().GetNativeWindow());
+    auto state = glfwGetKey(window, static_cast<int32_t>(key));
+    return state == GLFW_PRESS;
+}
 
-	bool Input::IsMouseButtonPressed(const MouseCode button)
-	{
-		auto* window = static_cast<GLFWwindow*>(Application::Get().GetWindow().GetNativeWindow());
-		auto state = glfwGetMouseButton(window, static_cast<int32_t>(button));
-		return state == GLFW_PRESS;
-	}
+bool Input::IsMouseButtonPressed(const MouseCode button)
+{
+    auto* window = static_cast<GLFWwindow*>(Application::Get().GetWindow().GetNativeWindow());
+    auto state = glfwGetMouseButton(window, static_cast<int32_t>(button));
+    return state == GLFW_PRESS;
+}
 
-	glm::vec2 Input::GetMousePosition()
-	{
-		auto* window = static_cast<GLFWwindow*>(Application::Get().GetWindow().GetNativeWindow());
-		double xpos, ypos;
-		glfwGetCursorPos(window, &xpos, &ypos);
+glm::vec2 Input::GetMousePosition()
+{
+    auto* window = static_cast<GLFWwindow*>(Application::Get().GetWindow().GetNativeWindow());
+    double xpos, ypos;
+    glfwGetCursorPos(window, &xpos, &ypos);
 
-		return { (float)xpos, (float)ypos };
-	}
+    return { (float)xpos, (float)ypos };
+}
 
-	float Input::GetMouseX()
-	{
-		return GetMousePosition().x;
-	}
+float Input::GetMouseX()
+{
+    return GetMousePosition().x;
+}
 
-	float Input::GetMouseY()
-	{
-		return GetMousePosition().y;
-	}
+float Input::GetMouseY()
+{
+    return GetMousePosition().y;
+}
 
 }