+++ Tabby/src/Tabby/Renderer/OrthographicCamera.cpp
@@ -9,14 +9,14 @@ OrthographicCamera::OrthographicCamera(float left, float right, float bottom, fl
     : m_ProjectionMatrix(glm::ortho(left, right, bottom, top, -1.0f, 1.0f))
     , m_ViewMatrix(1.0f)
 {
-    // TB_PROFILE_FUNCTION();
+    TB_PROFILE_SCOPE();
 
     m_ViewProjectionMatrix = m_ProjectionMatrix * m_ViewMatrix;
 }
 
 void OrthographicCamera::SetProjection(float left, float right, float bottom, float top)
 {
-    // TB_PROFILE_FUNCTION();
+    TB_PROFILE_SCOPE();
 
     m_ProjectionMatrix = glm::ortho(left, right, bottom, top, -1.0f, 1.0f);
     m_ViewProjectionMatrix = m_ProjectionMatrix * m_ViewMatrix;
@@ -24,7 +24,7 @@ void OrthographicCamera::SetProjection(float left, float right, float bottom, fl
 
 void OrthographicCamera::RecalculateViewMatrix()
 {
-    // TB_PROFILE_FUNCTION();
+    TB_PROFILE_SCOPE();
 
     glm::mat4 transform = glm::translate(glm::mat4(1.0f), m_Position) * glm::rotate(glm::mat4(1.0f), glm::radians(m_Rotation), glm::vec3(0, 0, 1));